// src/features/auth/schema.ts
import z from "zod";



export const loginSchema = z.object({
    email:z.string().email(),
    password: z.string().min(1,"Required"),
});

export const registerSchema = z.object({
    email: z.string().email(),
    name: z.string().trim().min(1,"Required"),
    password: z.string().min(1,"Required"),
});